Address WebUI projection review
